Uczesanie i uzupłnienie kodu z zajęć. Testy jednostkowe, wyjątki.

+++ zajecia_13/RestApiWithTests/src/main/java/org/example/mvc_demo/service/UserServiceImpl.java
@@ -1,6 +1,7 @@
 package org.example.mvc_demo.service;
 
 import org.example.mvc_demo.exception.NoSuchUserException;
+import org.example.mvc_demo.exception.UserAlreadyExistsException;
 import org.example.mvc_demo.model.User;
 import org.example.mvc_demo.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -8,26 +9,25 @@ import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.NoSuchElementException;
+import java.util.Optional;
 
+/**
+ * Serwis odpowiedzialny za zarządzanie użytkownikami.
+ */
 @Service
 public class UserServiceImpl implements UserService{
-
-    // private List<User> users = new ArrayList<User>();
 
     @Autowired
     private UserRepository userRepository;
 
     public UserServiceImpl()
     {
-      //  users.add(new User(1, "Marek", "abc@abc.ab"));
-      //  users.add(new User(2, "Janek", "abcd@a.b"));
-      //  users.add(new User(3, "Ala", "zzz@aaa.dd"));
     }
 
     @Override
     public List<User> findAll() {
         return userRepository.findAll();
-        // return users;
     }
 
 
@@ -36,7 +36,7 @@ public class UserServiceImpl implements UserService{
      *
      * @param id identyfikator użytkownika
      * @return znaleziony użytkownik
-
+     * @throws NoSuchElementException gdy użytkownik nie istnieje
      */
     @Override
     public User findById(int id) {
@@ -44,34 +44,49 @@ public class UserServiceImpl implements UserService{
         User user = userRepository.findById(id).orElse(null);
         if(user==null)
         {
-            throw new NoSuchUserException("Nie znalezion usera" + id);
+            throw new NoSuchUserException("Nie znaleziono usera " + id);
         }
         return user;
-        //return users.stream().filter(user -> user.getId() == id).findFirst().orElse(null);
     }
+
 
     @Override
     public void updateUser(User user) {
-        //
+        Optional<User> existingUser = userRepository.findById(user.getId());
+        if (!existingUser.isPresent()) {
+            throw new NoSuchUserException("No Such Customer exists: " + user.getId());
+        } else {
+            existingUser.get().setName(user.getName());
+            userRepository.save(existingUser.get());
+        }
     }
+
 
     @Override
     public void deleteUser(int id) {
-
+        Optional<User> existingUser = userRepository.findById(id);
+        if (!existingUser.isPresent()) {
+            throw new NoSuchUserException("No Such User exists: " + id);
+        }
         userRepository.deleteById(id);
-        //users.removeIf(user -> user.getId() == id);
     }
 
     @Override
     public void addUser(User user) {
-
+        Optional<User> existingUser = userRepository.findUsersByEmail(user.getEmail());
+        if (existingUser.isPresent()) {
+            throw new UserAlreadyExistsException("User with this email already exists");
+        }
         userRepository.save(user);
-        //users.add(user);
     }
 
     @Override
     public void deleteAllUsers() {
         userRepository.deleteAll();
-        //users.clear();
+    }
+
+    @Override
+    public User findByEmail(String email) {
+        return userRepository.findUsersByEmail(email).orElseThrow( () -> new NoSuchUserException("Nie znaleziono usera " + email));
     }
 }
